requirements.txt
autopep8==1.5.6
bidict==0.21.2
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
firebase-admin==4.5.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.51.1
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.1
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.24.2
opencv-python==4.7.0.68
packaging==20.9
pandas==1.3.4
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.15.0
sklearn==0.0
threadpoolctl==3.0.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
